ChatInput 별도 컴포넌트로 꺼냄
