+++ components/chat/ActiveChat.tsx
@@ -1,8 +1,6 @@
-import { useRecoilValue, useSetRecoilState } from 'recoil';
+import { useRecoilValue } from 'recoil';
 import styled from 'styled-components';
 import ChatContent from './ChatContent';
-import ImageOutlinedIcon from '@mui/icons-material/ImageOutlined';
-import SendRoundedIcon from '@mui/icons-material/SendRounded';
 import SearchIcon from '@mui/icons-material/Search';
 import VideocamRoundedIcon from '@mui/icons-material/VideocamRounded';
 import {
@@ -10,57 +8,15 @@ import {
   activeChatRoomState,
 } from '../../recoil/selectors';
 import ProfileImage from '../common/ProfileImage';
-import useInput from '../../hooks/useInput';
-import socket from '../../library/socket';
-import { SocketEvents } from '../../library/socket.events.enum';
-import { useEffect, useRef, useState } from 'react';
-import { chatState } from '../../recoil/atoms';
-import { Message } from '../../interfaces/chat.interface';
+import { useEffect, useRef } from 'react';
 import usePeerConnection from '../../hooks/usePeerConnection';
+import ChatInput from './ChatInput';
 
 export default function ActiveChat() {
   const activePartner = useRecoilValue(activeChatPartnerState);
   const activeChatRoom = useRecoilValue(activeChatRoomState);
-  const setChatData = useSetRecoilState(chatState);
-  const [isPending, setIsPending] = useState(false);
-  const [value, onChangeInputText, setInputText] = useInput();
   const messageContainerRef = useRef<HTMLDivElement | null>(null);
-  const SendBtnColor = value.length ? '#8083FF' : '#727272';
   const { handleRequestCall } = usePeerConnection();
-
-  const handleInputKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (!value || isPending) {
-      return;
-    }
-    if (e.code === 'Enter') {
-      setIsPending(true);
-      const payload = {
-        chatRoom_id: activeChatRoom?._id,
-        type: 'text',
-        content: value,
-      };
-      socket.emit(SocketEvents.Message, payload, (message: Message) => {
-        setInputText('');
-        setIsPending(false);
-        setChatData((prev) => {
-          const filteredChatRoom = prev.filter(
-            (chatRoom) => chatRoom._id !== message.chatRoom_id,
-          );
-          const target = prev.find(
-            (chatRoom) => chatRoom._id === message.chatRoom_id,
-          );
-          if (target) {
-            const targetRoomMessages = [...target.messages, message];
-            return [
-              { ...target, messages: targetRoomMessages },
-              ...filteredChatRoom,
-            ];
-          }
-          return prev;
-        });
-      });
-    }
-  };
 
   useEffect(() => {
     if (messageContainerRef.current) {
@@ -107,16 +63,7 @@ export default function ActiveChat() {
         />
         <div ref={messageContainerRef}> </div>
       </MessageContainer>
-      <MessageInputContainer>
-        <ImageOutlinedIcon sx={{ color: '#727272', fontSize: 23 }} />
-        <input
-          value={value}
-          onChange={onChangeInputText}
-          placeholder="Your messages..."
-          onKeyDown={handleInputKeyDown}
-        />
-        <SendRoundedIcon sx={{ color: SendBtnColor, fontSize: 23 }} />
-      </MessageInputContainer>
+      <ChatInput activeChatRoom={activeChatRoom} />
     </Container>
   );
 }
@@ -180,35 +127,5 @@ const VideoButton = styled.button`
   svg {
     fill: ${({ theme }) => theme.grayColor};
     aspect-ratio: 1;
-  }
-`;
-
-const MessageInputContainer = styled.div`
-  height: 7rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin: 1rem;
-  height: 4.5rem;
-  padding: 1.5rem;
-  background-color: #242526;
-  border-radius: 1rem;
-  input {
-    width: 100%;
-    background-color: inherit;
-    margin-left: 1rem;
-    color: ${({ theme }) => theme.fontColor.contentColor};
-    ::placeholder {
-      color: ${({ theme }) => theme.fontColor.contentColor};
-    }
-    &:focus {
-      outline: none;
-    }
-  }
-  svg {
-    &:last-child {
-      transform: rotateZ(-45deg);
-      margin-bottom: 0.5rem;
-    }
   }
 `;
